+++ src/fhir/models/diagnostic-report.test.ts
@@ -1,15 +1,23 @@
-import { Coding, DiagnosticReport, Observation } from "fhir/r4";
+import { faker } from "@faker-js/faker";
+import { Coding, DiagnosticReport, Observation, Reference } from "fhir/r4";
 import * as dr from "./diagnostic-report";
+import { ObservationModel } from "./observation";
+import { SYSTEM_CPT, SYSTEM_LOINC } from "../system-urls";
 
-const syntheticObservation = (): Observation => ({
+const syntheticObservation = (coding: Coding[], text?: string): Observation => ({
+  id: faker.datatype.uuid(),
   resourceType: "Observation",
-  status: "unknown",
-  code: {},
+  status: "final",
+  code: {
+    coding,
+    text,
+  },
 });
 
-const syntheticDiagnosticReport = (codings: Coding[], text?: string): DiagnosticReport => ({
+const syntheticDiagnosticReport = (coding: Coding[], text?: string): DiagnosticReport => ({
+  id: faker.datatype.uuid(),
   code: {
-    coding: codings,
+    coding,
     text,
   },
   resourceType: "DiagnosticReport",
@@ -41,6 +49,145 @@ describe("Diagnostic Report Model Tests", () => {
     });
   });
 
+  describe("constructor", () => {
+    let d: DiagnosticReport;
+    let o1: ObservationModel;
+    let o2: ObservationModel;
+    let o3: ObservationModel;
+    let o4: ObservationModel;
+
+    beforeEach(() => {
+      d = syntheticDiagnosticReport([]);
+      o1 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "17856-6", // a1c
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+      o2 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "4548-4", // a1c
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+      o3 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "2345-7", // glucose
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+      o4 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "2345-7", // glucose
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+    });
+
+    test("no trends", () => {
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, []);
+      expect(model.hasTrends).toBe(false);
+      expect(model.observations).toEqual([]);
+    });
+
+    test("unassociated trends", () => {
+      d.result = [o1 as Reference, o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o1, o3]);
+      expect(model.hasTrends).toBe(false);
+      expect(model.observations).toHaveLength(2);
+      expect(model.observations[0].trends).toEqual([o1]);
+      expect(model.observations[1].trends).toEqual([o3]);
+    });
+
+    test("associated trends", () => {
+      d.result = [o1 as Reference, o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o1, o2, o3, o4]);
+      expect(model.hasTrends).toBe(true);
+      expect(model.observations).toHaveLength(2);
+      expect(model.observations[0].trends).toEqual([o1, o2]);
+      expect(model.observations[1].trends).toEqual([o3, o4]);
+    });
+
+    test("associated trends w/ incorrectly coded glucose (LOINC)", () => {
+      d = syntheticDiagnosticReport([
+        {
+          code: "4548-4",
+          system: SYSTEM_LOINC,
+        },
+      ]);
+      o3 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "2345-7", // glucose
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+      d.result = [o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o3, o4]);
+      expect(model.hasTrends).toBe(false);
+      expect(model.observations).toHaveLength(1);
+      expect(model.observations[0].trends).toHaveLength(0);
+    });
+
+    test("associated trends w/ incorrectly coded glucose (CPT)", () => {
+      d = syntheticDiagnosticReport([
+        {
+          code: "83036",
+          system: SYSTEM_CPT,
+        },
+      ]);
+      o3 = new ObservationModel(
+        syntheticObservation([
+          {
+            code: "2345-7", // glucose
+            system: SYSTEM_LOINC,
+          },
+        ])
+      );
+      d.result = [o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o3, o4]);
+      expect(model.hasTrends).toBe(false);
+      expect(model.observations).toHaveLength(1);
+      expect(model.observations[0].trends).toHaveLength(0);
+    });
+
+    test("associated trends w/ incorrectly coded glucose (display)", () => {
+      d = syntheticDiagnosticReport([
+        {
+          display: "a1c",
+        },
+      ]);
+      d.result = [o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o3, o4]);
+      expect(model.hasTrends).toBe(false);
+      expect(model.observations).toHaveLength(1);
+      expect(model.observations[0].trends).toHaveLength(0);
+    });
+
+    test("associated trends w/ incorrectly and correctly coded glucose", () => {
+      d = syntheticDiagnosticReport([
+        {
+          display: "a1c",
+        },
+      ]);
+      d.result = [o1 as Reference, o3 as Reference];
+      const model = new dr.DiagnosticReportModel(d, undefined, undefined, [o1, o2, o3, o4]);
+      expect(model.hasTrends).toBe(true);
+      expect(model.observations).toHaveLength(2);
+      expect(model.observations[0].trends).toEqual([o1, o2]);
+      expect(model.observations[1].trends).toHaveLength(0);
+    });
+  });
+
   describe("infer start date", () => {
     interface TestType {
       obs: (Observation | undefined)[] | undefined;
@@ -55,19 +202,19 @@ describe("Diagnostic Report Model Tests", () => {
       {
         expectedValue: "04/01/2022",
         obs: [
-          set("effectiveDateTime", syntheticObservation(), "2022-04-02"),
-          set("effectiveDateTime", syntheticObservation(), "2022-04-01"),
+          set("effectiveDateTime", syntheticObservation([]), "2022-04-02"),
+          set("effectiveDateTime", syntheticObservation([]), "2022-04-01"),
         ],
       },
       {
         expectedValue: "04/01/2022",
-        obs: [undefined, set("effectiveDateTime", syntheticObservation(), "2022-04-01")],
+        obs: [undefined, set("effectiveDateTime", syntheticObservation([]), "2022-04-01")],
       },
       {
         expectedValue: "04/01/2022",
         obs: [
           undefined,
-          set("effectivePeriod", syntheticObservation(), { start: "2022-04-01", end: undefined }),
+          set("effectivePeriod", syntheticObservation([]), { start: "2022-04-01", end: undefined }),
         ],
       },
     ];