+++ src/services/fqs/queries/diagnostic-reports-query.ts
@@ -102,9 +102,7 @@ export const diagnosticReportQuery = gql`
           }
           contained {
             resource {
-              ... on Observation {
-                ...Observation
-              }
+              ...Observation
             }
           }
         }